Added the instuction functions that take the program and expand instructions and their arguments. Also added functions that replace the labels with addresses and replace everything with bytes. Just to see it working also added the function that runs all stages and produces a list of bytes to be written.
Currently these are all only implemented to properly assemble the test
file. I ran the result with the VM and it produces the expected output.
Getting the rest of the ops working is a matter of adding some extra
cases to existing functions and some keys to sets. This can be done for
a few ops at a time along with some files that implement different
algorithms or programs to test out the assembler and more VM
functionality.