Fixed various errors due to moving and renaming things without restarting the repl.
Added a get label function because a label can make
it through to the replace labels pass and then cause issues when we try
to replace it. If it doesn't exist using get returns nil and then we
might try to convert that to a number/word which will error, but not in
an intuitive way for a user. Now it will explicitly say the label
doesn't exits.

Setup the assemble function and got it running. It is now possible to
pass it the program file and get a byte-array back. I have not yet
tested writing this to a file and running it on the vm because there are
some small issues I noticed in the vm. Notably the branch checks the top
for true and then gets the address. This need to be the other way around
so that pushing the address is in an intuitive place.