Wrote another sample problem that solves an actual Kattis problem when assembled to C. In the process discovered some small issues.
The change I made to the replace push was not quite correct. I both
forgot to pass the proper map for the num->bytes args and didn't
actually call the function where it was supposed to replace others.

Had to add to the VM to allow printmem to take a flag like load and
store to use a full address rather than an offset to fmp. Then added a
:prnmem-lb op to the vm to set that up.

Added some error handling to num->bytes because there was an issue
passing it a keyword instead of a map, and it just returned nil, which
was not helpful.

The labeling process was not counting 2 bytes when a :dpush happened.
This is super important because otherwise the program counter and labels
can go off by one and not work. It was not an issue in the first test
because it didn't use dwords. It is fixed for :dpush and :fpush, but
possibly there are other areas, like static that it might cause issues.