Added the start of what could be a standard lib. This is to replace pseudo and compund ops. Instead short groups of ops that might be handy, but are a bit long can be made into stdlib subroutines. The main thing here is that anything that would be 3 or fewer ops should not be a subroutine because it would add more ops than is worth it. I mean we are adding a :push label :call and a :ret to wrap around what is being replaced. The subroutine call is not really expensive, but if something trivial is added to the stdlib and only called once in a program it is a waste.
Note that the purpose of the stdlib will not be to make programming in
the assembly language really easy, it is still an assembly language.
Instead it should be to create subroutines for common VM like tasks that
are long or error prone. The current subroutines added are not really
what I have in mind, but I needed something and they are easy to create.