Wrote another simple program to test the assembler. The tests show that the uberjar works as expected, though updated the readme to note that the jar has a slow startup.
Noticed some issues with the double word versions of comparissons. They
return double words and I think they need to be changed to return words.
The problem is that branch pulls only words so the stack gets out of
wack. It is possible to swap and pop the double word to turn it into a
word, as done in the coldputer.lta test project.