Added support for builtin macros/psuedo instructions and tested some simple ones out in the existing test programs.
I thought that the
stdlib would be enough, but for things 3 ops or less the use of a
subroutine is not ideal as it does not reduce typed code and adds to
program size. So macros that are just replaced by their ops are better
for small conviniences. For larger things it is not as useful as a
subroutine includes the called code only once and the call is shorter
than the original set of ops.

One of the nice things about macros is that they can make some small
tasks a little more obvious as to what is happening. For example :->word
is much more descriptive than :swap :pop when turning a double word into
a word. The other use is to make up for instruction limitations in the
VM. The word conversion example is a good one, as the vm returns a
double size word currently for double word comparisson, but branch takes
a word for the result value. This particular thing should probably be
changed in the VM as all boolean like values should be a word. However,
having macros can make it much easier to deal with. For example a macro
:deq-w could be added that calls :deq :swap :pop to return a word from
deq. This obviously adds more instructions than if the vm did it, but it
is eaier to change than messing with how the VM works, especially when
the VM gets to a point where changing ops would break existing programs.

There is no support for user macros yet, however it would not be too
hard to add a map to the program data that holds them. Then during
replacement we can add another almost identical line to the existing
macro replacement that will check if something is a user macro.