Added user defined macros. User macros cannot contain labels, though it is not checked yet. The idea is that they are just operations and literals, though labels could be added, but it would change where macro substitution is done, and honestly the only place it is really useful is to shorten a call statement that is used a lot. Since macros are designed to be short substitutions and not involved, if labels need to be used then a subroutine is the best way to do it.
User macro syntax is (macro :!name & ops). The name does not have to be
a keyword or use the !, but it seems like a good convention that will
keep them from getting mixed up by the programmer or the assembler.
Also, user macros are checked before builtin macros so it is possible to
shadow builtin macros in a program, or a module.