Moved the kattis tests around, and deleted the original simple test and module. Added a folder to put non stdlib modules in for use in programs or to test out the vm more. Put a few of the latest macros into the macro map.
Most importantly setup a test file that uses the clojure test suite to
test these kattis programs and any others that might be useful for
testing the vm with more complicated programs. For each tests a setup
function takes a .lta file and assembles it into a temporary standalone
.c file. That file is compiled and then a function that will execute
the resulting program on a list of lines as input. The function will
return the oputput of the program with trailing whitespace trimmed. The
only downside for the vm is that if it reads a line there must be a
newline at the end, becuase it joins the lines and that won't add a
newline to the final string. So if it should end with a newline it needs
to be added explicitly. However, this addition makes it way easier to
test both the vm and the assembler, and any test programs when they are
being created. Changes can be made and the results quickly and easily
seen.