Modified the main routine to catch assembly errors for the standalone c file option and print the error message. This is useful as caught exceptions in clojure can be very long and gross. Error messages for the assembly are not as informative as they could be, but it is assembly and they are enough for now, especially with small programs or subroutines.
Updated the VM testing file to better handle errors assembling and
compiling so that there are not long unusable error messages when
running tests. Now the general error message is passed as the result of
the execution funtion. This ensures test failure and gives an idea of
the reason. The full reason can be seen by scrolling up. Since these are
standalone programs being run, complex errors or compile errors should
be diagnosed from outside of the tests, so no detailed information is
given for them.