Added stepping to the debug mode. It pauses and asks for a step after printing details. To just advance one step press enter without any input. To have it execute a number of operations before pausing again enter a positive number of steps to make and press enter.
Note that if input is required to be entered for the program after the
step without prompt it can seem like pressing enter for the step didn't
work, but it has an the program is waiting for program input.

Also, the first step is to always skip the initial jump instruction that
skips the static data, as it is not part of the user program.