Fixed some small issues with the debugging. It now has the guaranteed behaviour of skipping one step after any read that does not remove the newline. This is preferable to alternatives that either require a 0 to be input to move to the next step, or some non standard way to try and clear the input buffer.
It should be noted that the during debug mode if a read is made, even if
the input is expected to be a bunch of number on a line or something,
the user should press enter after the next unit of characters that are
expected to be read by the next call. So if you have wread add a single
number not a list of them, the program will loop around to the next one.