Changed the VM debugging to use break points instead of step counts. A :BREAK op can now be added to the program that will stop the debugger if it is looking for one. The step now just steps if return is pressed or if any input is given it will run until the next break. Note that reads that do not give a prompt will pause for input even between breaks, which can be confusing if trying to get to a far away break.
Added debugging behaviour for individual reads to eat a char,
The newline after entering the element. This means that no steps are
automatically taken, but means that read loops that take chars cannot be
given input like they would in a program, i.e. steve, but instead must
be given a single char and then press return till the next is asked for,
i.e. s<CR>t<CR>...